+++ packages/fedify/src/testing/context.ts
@@ -51,6 +51,8 @@ export function createContext<TContextData>(
     lookupWebFinger,
     sendActivity,
     routeActivity,
+    enqueueTask,
+    enqueueTaskMany,
   } = values;
   function throwRouterError(): URL {
     throw new RouterError("Not implemented");
@@ -111,6 +113,12 @@ export function createContext<TContextData>(
       throw new Error("Not implemented");
     }),
     routeActivity: routeActivity ?? ((_params) => {
+      throw new Error("Not implemented");
+    }),
+    enqueueTask: enqueueTask ?? ((_task, _data, _options) => {
+      throw new Error("Not implemented");
+    }),
+    enqueueTaskMany: enqueueTaskMany ?? ((_task, _payloads, _options) => {
       throw new Error("Not implemented");
     }),
   };